rectangels
Bugfix/116 heatmap toggle
